requirements.txt
# API и HTTP
requests>=2.31.0
python-dotenv>=1.0.0

# Groq для генерации текста (используем более новую версию для совместимости)
groq>=0.11.0

# Telegram бот (используем более новую версию)
python-telegram-bot>=21.0

# Планировщик задач
schedule>=1.2.0

# YouTube API
google-api-python-client>=2.108.0
google-auth>=2.25.2
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Веб-интерфейс (UI) - используем более новую версию
flet>=0.23.0

# Утилиты
pyyaml>=6.0.1
psutil>=5.9.0

# Open-source видео генератор (альтернатива HeyGen)
moviepy>=1.0.3  # Создание видео
gtts>=2.5.0  # Google Text-to-Speech (бесплатно, базовое качество)
pillow>=10.1.0  # Работа с изображениями

# Для улучшенного голоса (опционально, РЕКОМЕНДУЕТСЯ)
# TTS>=0.22.0  # Coqui TTS - высокое качество для русского языка
# Раскомментируйте для премиум голоса:
# pip install TTS

# Опциональные зависимости
# opencv-python>=4.8.1  # для продвинутой обработки видео
